pyproject.toml
[tool.poetry]
name = "fedlib"
version = "0.0.12345"
description = "A scalable federated learning library based on PyTorch and Ray."
authors = ["Shenghui Li <shenghui.li@it.uu.se>"]
license = "Apache-2.0"
readme = "README.rst"
homepage = "https://github.com/fedlib/fedlib"
repository = "https://github.com/fedlib/fedlib"
documentation = "https://fedlib.readthedocs.io"
keywords = ["deep-learning", "pytorch", "federated-learning"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3 :: Only",
]

[tool.poetry.dependencies]
python = "^3.9"
torchvision = "*"
torchaudio = "*"
ray = { version = "^2.30.0", extras = ["rllib", "tune"] }
requests = "^2.27.1"
setuptools = "*"
tqdm = "*"
wandb = "^0.17"
dm_tree = "*"
torchmetrics = "*"

# [tool.poetry.dev-dependencies]
pytest-xdist = "*"
pre-commit = "*"
scikit-learn = "*"
yapf = "*"
isort = "*"
codespell = "*"
black = "*"
pylint = "*"
pytest = "*"
#[tool.poetry.scripts]
#blades = "blades.train:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
skip = [".gitignore", "__init__.py"]
